+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/viewmodel/SubmissionFragmentModule.kt
@@ -8,8 +8,8 @@ import de.rki.coronawarnapp.ui.submission.fragment.SubmissionContactFragment
 import de.rki.coronawarnapp.ui.submission.fragment.SubmissionDispatcherFragment
 import de.rki.coronawarnapp.ui.submission.deletionwarning.SubmissionDeletionWarningFragment
 import de.rki.coronawarnapp.ui.submission.deletionwarning.SubmissionDeletionWarningModule
-import de.rki.coronawarnapp.ui.submission.greencertificate.RequestCovidCertificateFragment
-import de.rki.coronawarnapp.ui.submission.greencertificate.RequestCovidCertificateFragmentModule
+import de.rki.coronawarnapp.ui.submission.covidcertificate.RequestCovidCertificateFragment
+import de.rki.coronawarnapp.ui.submission.covidcertificate.RequestCovidCertificateFragmentModule
 import de.rki.coronawarnapp.ui.submission.qrcode.consent.SubmissionConsentFragment
 import de.rki.coronawarnapp.ui.submission.qrcode.consent.SubmissionConsentModule
 import de.rki.coronawarnapp.ui.submission.qrcode.scan.SubmissionQRCodeScanFragment
@@ -102,7 +102,7 @@ internal abstract class SubmissionFragmentModule {
     abstract fun submissionTestResultNoConsentScreen(): SubmissionTestResultNoConsentFragment
 
     @ContributesAndroidInjector(modules = [SubmissionResultPositiveOtherWarningNoConsentModule::class])
-    abstract fun SubmissionResultPositiveOtherWarningNoConsentScreen(): SubmissionResultPositiveOtherWarningNoConsentFragment
+    abstract fun submissionResultPositiveOtherWarningNoConsentScreen(): SubmissionResultPositiveOtherWarningNoConsentFragment
 
     @ContributesAndroidInjector(modules = [SubmissionDeletionWarningModule::class])
     abstract fun submissionDeletionWarningScreen(): SubmissionDeletionWarningFragment
@@ -114,5 +114,5 @@ internal abstract class SubmissionFragmentModule {
     abstract fun submissionTestResultKeysSharedScreen(): SubmissionTestResultKeysSharedFragment
 
     @ContributesAndroidInjector(modules = [RequestCovidCertificateFragmentModule::class])
-    abstract fun requestGreenCertificateFragment(): RequestCovidCertificateFragment
+    abstract fun requestCovidCertificateFragment(): RequestCovidCertificateFragment
 }
