+++ Corona-Warn-App/src/main/res/layout/fragment_request_covid_certificate.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context="ui.submission.greencertificate.RequestCovidCertificateFragment">
+    tools:context="ui.submission.covidcertificate.RequestCovidCertificateFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
